requirements.txt
ipython==8.8.0
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
plotly==5.9.0
pyairtable==1.4.0
requests==2.28.1
seaborn==0.11.2
streamlit==1.16.0
wordcloud==1.8.2.2
